[core] Generalize Map::{add,remove}CustomLayer
